requirements.txt
langchain>=0.0.350
openai>=1.3.7
python-dotenv>=1.0.0
stability-sdk>=0.8.5
moviepy>=1.0.3
gtts>=2.4.0
pydantic>=2.5.2
python-magic>=0.4.27
Pillow>=10.1.0
tiktoken>=0.5.1
